+++ demos/src/Examples/MultiMention/Vue/index.spec.js
@@ -1,0 +1,7 @@
+context('/src/Nodes/Mention/Vue/', () => {
+  beforeEach(() => {
+    cy.visit('/src/Nodes/Mention/Vue/')
+  })
+
+  // TODO: Write tests
+})