+++ packages/extension-drag-handle-react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @tiptap/extension-drag-handle-react
 
+## 2.23.1
+
+### Patch Changes
+
+- 5b3d723: Fix a bug that causes rerenders on the react drag handle breaking its tippy connection
+
+## 2.23.0
+
+## 2.22.3
+
+## 2.22.2
+
+## 2.22.1
+
 ## 2.22.0
 
 ### Minor Changes