+++ packages/extension-mathematics/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @tiptap/extension-mathematics
 
+## 2.23.1
+
+## 2.23.0
+
+## 2.22.3
+
+## 2.22.2
+
+## 2.22.1
+
 ## 2.22.0
 
 ### Minor Changes